pyproject.toml
[tool.poetry]
name = "esbook"
version = "0.1.0"
description = "The REST API platform that provides search within millions of books."
authors = ["gixproject <viacheslavlab@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.2"
flask = "^1.1.2"
elasticsearch = "^7.8.0"
Flask-SQLAlchemy = "^2.4.3"
Flask-Script = "^2.0.6"
Flask-Migrate = "^2.5.3"
psycopg2 = "^2.8.5"
passlib = "^1.7.2"
psycopg2-binary = "^2.8.5"
flask-restplus = "^0.13.0"
Werkzeug = "0.16.1"
marshmallow = "^3.8.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.23.1"
Faker = "^4.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.3"
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
